add max speed and arming move threshold

+++ OnPod/FlightComputer/src/Threads/CanBusThread.cpp
@@ -159,11 +159,11 @@ void startInverterBroadcast(int bcmSocket){
     }
 }
 
-void setInverterTorque(int torque, int bcmSocket){
+void setInverterTorque(int32_t torque, int bcmSocket){
     struct broadcastManagerConfig msg = {};
     torque *=10;
-    int32_t highByte = torque/256;
-    int32_t lowByte = torque%256;
+    uint16_t highByte = torque/256;
+    uint16_t lowByte = torque%256;
     msg.msg_head.opcode  = TX_SETUP;
     msg.msg_head.can_id  = 0;
     msg.msg_head.flags   = SETTIMER | STARTTIMER;
@@ -256,7 +256,7 @@ int canNetworkThread(TelemetryManager Pod){
         return -1;
     }
     try{
-       //  startInverterBroadcast(canSockBcm);
+       ;//  startInverterBroadcast(canSockBcm);
     }
     catch (const std::runtime_error &error){
         LOG(INFO) << error.what();
